% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APY.R
\name{APY}
\alias{APY}
\title{Convert CMT to APY}
\usage{
APY(i, percent = FALSE)
}
\arguments{
\item{i}{vector of CMT yields to be converted}

\item{percent}{boolean, if TRUE, i should be divided by 100
(for example 7.87 becomes 0.0787)}
}
\value{
APY yield(s) as a decimal (not as a percent)
}
\description{
Convert Constant-Maturity Treasury (CMT) yields to
Annualized Percentage Yields (APY)
}
\examples{
APY(0.0800) #> [1] 0.0816

CMT = c(7.87, 7.92, 7.91, 7.92, 7.92, 7.92)
APY(CMT, percent=TRUE)
#> [1] 0.08024842 0.08076816 0.08066420 0.08076816 0.08076816 0.08076816

}
\references{
Interest Rates - Frequently Asked Questions

\url{https://www.treasury.gov/resource-center/faqs/Interest-Rates/Pages/faq.aspx#2}
}

