% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{CAGRd}
\alias{CAGRd}
\title{Compound Annual Growth Rate, discrete}
\usage{
CAGRd(Starting_value, Ending_Value, Number_of_years)
}
\arguments{
\item{Starting_value}{the price at the beginning of the period}

\item{Ending_Value}{the price at the end of the period}

\item{Number_of_years}{the length of the period in (fractional) years}
}
\value{
the compounded rate of return over discrete intervals
}
\description{
Compound Annual Growth Rate, discrete
}
\note{
see \emph{r_continuous} and \emph{r_discrete}
}
\examples{
# Wikipedia
# https://en.wikipedia.org/wiki/Compound_annual_growth_rate#Example
Starting_value  <- 9000
Ending_Value    <- 13000
Number_of_years <- 3
ans <- CAGRd(Starting_value, Ending_Value, Number_of_years)*100
writeLines(paste0(round(ans,0),"\%"))

}

