% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{CallParity}
\alias{CallParity}
\title{Call prices derived from put-call parity}
\usage{
CallParity(Stock, Exercise, Time, Interest, Yield, Put_price)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{Put_price}{the price of the put option to convert}
}
\description{
Call prices derived from put-call parity
}
\examples{
# Hull 7th edition Ch 17 P 357
library(ustreasuries)
Stock    <- 49
Exercise <- 50
Time     <- 20/52
Interest <- 0.05
Yield    <- 0#0.13
sigma    <- 0.20

EC = EuroCall(Stock, Exercise, Time, Interest, Yield, sigma)
EP = EuroPut(Stock, Exercise, Time, Interest, Yield, sigma)

PC = CallParity(Stock, Exercise, Time, Interest, Yield, EP)
PP = PutParity(Stock, Exercise, Time, Interest, Yield, EC)

writeLines(paste0("European Call Price: ", EC, "\\n",
                  "Call Parity Price:   ", PC, "\\n",
                  "Difference:          ", EC-PC, "\\n\\n",

                  "European Put Price:  ", EP, "\\n",
                  "Put Parity Price:    ", PP, "\\n",
                  "Difference:          ", EP-PP))
}

