% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{continuously_compounded_rate}
\alias{continuously_compounded_rate}
\title{Calculate continuously-compounded annualized rate of interest
(CAGR or r(t, T)) from a discount factor (Z(t, T))}
\usage{
continuously_compounded_rate(d_f, years)
}
\arguments{
\item{d_f}{discount factor (Z(t, T))}

\item{years}{number of (fractionalized) years: period that the discount factor
effects}
}
\value{
r(t, T)
}
\description{
\itemize{
     \item See CAGR() to calculate continuously-compounded annualized rate of
     interest (r(t, T)) from starting and ending asset values
     \item See r_continuous() to calculate r(t, T) from discrete-period
     interest rates
     \item See discount_factor() to calculate  the discount factor (Z(t, T))
     from a continuously-compounded annualized rate of interest (r(t,T))
 }
}
\note{
PV = Z(t,T) * $1
}
\examples{
# Veronesi Ch2 P38
d_f   <- 0.9953996
years <- 13 / 52
continuously_compounded_rate(d_f, years) # 0.018444

}
\references{
Veronesi Ch2 P29-38
}

