% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{DeltaCall}
\alias{DeltaCall}
\title{Delta of a European Call Option}
\usage{
DeltaCall(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Delta of the call option
}
\description{
Given a change in asset price, DeltaCall describes the amount by which the
call option price changes.
}
\details{
\code{Change_in_Call-Option_Price = DeltaCall * Change_in_Asset_Price}
}
\examples{
# Hull, 7th edition Ch 17 p 363
library(ustreasuries)
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

ans <- DeltaCall(Stock, Exercise, Time, Interest, Yield, sigma)

writeLines(paste0("Delta call: when the asset price changes by Delta_S,\\n",
                  "                the option price changes by Delta_S*",round(ans, 3)))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

