% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{DeltaPut}
\alias{DeltaPut}
\title{Delta of a European Put Option}
\usage{
DeltaPut(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Delta of the put option
}
\description{
Given a change in asset price, DeltaPut describes the amount by which the
put option price changes.
}
\details{
\code{Change_in_Put-Option_Price = DeltaPut * Change_in_Asset_Price}
}
\examples{
# Hull, 7th edition Ch 17 p 362,3
library(ustreasuries)
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

dcall <- DeltaCall(Stock, Exercise, Time, Interest, Yield, sigma)
dput  <- DeltaPut(Stock, Exercise, Time, Interest, Yield, sigma)

writeLines(paste0("Delta put: when the asset price changes by Delta_S,\\n",
                  "               the option price changes by Delta_S*",round(dput, 3),
                  "\\nDelta put = Delta call - 1? ", dput == dcall-1))

}

