% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{discount_factor}
\alias{discount_factor}
\title{Calculate discount factor (Z(t, T)) from continuously-compounded annualized
rate of interest (CAGR or r(t, T))}
\usage{
discount_factor(cagr, years)
}
\arguments{
\item{cagr}{Compounded Annual Growth Rate: continuously-compounded annualized rate of
interest (r(t, T))}

\item{years}{number of (fractionalized) years: period that the discount factor
effects}
}
\value{
Z(t, T)
}
\description{
\itemize{
     \item See CAGR() to calculate continuously-compounded annualized rate of
     interest (r(t, T)) from starting and ending asset values
     \item See r_continuous() to calculate r(t, T) from discrete-period
     interest rates
     \item See continuously_compounded_rate() to calculate
     continuously-compounded annualized rate of interest (r(t,T))
     from a discount factor (Z(t, T))
 }
}
\note{
PV = Z(t,T) * $1
}
\examples{
# Veronesi Ch2 P38
cagr  <- 0.018444
years <- 13 / 52
discount_factor(cagr, years) # 0.9953996

}
\references{
Veronesi Ch2 P29-38
}

