% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{dOne}
\alias{dOne}
\title{The d1 component of the Black-Scholes-Merton formula}
\usage{
dOne(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
d1
}
\description{
The d1 component of the Black-Scholes-Merton formula
}
\examples{
# Hull 7th edition Ch 13 P 294
Stock     <- 42    # S_0
Exercise  <- 40    # K
Time      <- 0.50  # T
Interest  <- 0.10  # r
Yield     <- 0     # q
sigma     <- 0.20
ans <- dOne(Stock, Exercise, Time, Interest, Yield, sigma)
round(ans,4)

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull 7th edition Ch 13 P 291
}

