% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{en}
\alias{N}
\alias{en}
\title{N: the Standard Normal CDF}
\usage{
en(x)
}
\arguments{
\item{x}{real number}
}
\value{
N(x)
}
\description{
Equivalent to the R function \emph{pnorm}
}
\details{
Created numerically using the internal function \emph{phi}
}
\note{
same function provided by \emph{en, N}
}
\examples{
for (x in seq(from=-5, to=5, by=0.1)) {
   print(paste(x, en(x), pnorm(x), all.equal(en(x), pnorm(x), tolerance=0.01)))
    }

}

