% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{EuroCall}
\alias{EuroCall}
\title{European Call}
\usage{
EuroCall(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
the price of a European Call Option
}
\description{
A call option that can only be exercised at expiration
}
\examples{
# Hull 7th edition Ch 13 P 294
Stock     <- 42
Exercise  <- 40
Time      <- 0.50
Interest  <- 0.10
Yield     <- 0
sigma     <- 0.20
ans <- EuroCall(Stock, Exercise, Time, Interest, Yield, sigma)
round(ans,2)

}
\references{
Hull 7th edition Ch 13 P 291
}

