% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{EuroCallVol}
\alias{EuroCallVol}
\title{Implied Volatility for a European Call Option}
\usage{
EuroCallVol(Stock, Exercise, Time, Interest, Yield, Call_price)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{Call_price}{the price of the call option being valued}
}
\value{
The Implied Volatility of the call option
}
\description{
Implied Volatility for a European Call Option
}
\examples{
# Hull, 7th edition ch 13 p296
Stock      <- 21     # S_0
Exercise   <- 20     # K
Time       <- 0.25   # T
Interest   <- 0.10   # r
Yield      <- 0      # q
Call_price <- 1.875

callvol <- EuroCallVol(Stock, Exercise, Time, Interest, Yield, Call_price)
writeLines(paste0("Implied Call Volatility: ", round(callvol*100, 1), "\% per annum"))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition ch 13 p296-297
}

