% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{EuroPutVol}
\alias{EuroPutVol}
\title{Implied Volatility for a European Put Option}
\usage{
EuroPutVol(Stock, Exercise, Time, Interest, Yield, Put_price)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{Put_price}{the price of the put option being valued}
}
\value{
The Implied Volatility of the put option
}
\description{
Implied Volatility for a European Put Option
}
\examples{
# Hull, 7th edition ch 13 p296
Stock      <- 21     # S_0
Exercise   <- 20     # K
Time       <- 0.25   # T
Interest   <- 0.10   # r
Yield      <- 0      # q
Put_price <- 1.875

putvol <- EuroPutVol(Stock, Exercise, Time, Interest, Yield, Put_price)
writeLines(paste0("Implied Put Volatility: ", round(putvol*100, 1), "\% per annum"))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition ch 13 p296-297
}

