% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rates.R
\name{FedInvestData}
\alias{FedInvestData}
\title{Return treasury bond price data from 2010}
\usage{
FedInvestData()
}
\value{
data.frame containing the FedInvest data from 2010 to the most-recently
completed business day
}
\description{
The FedInvest site provides rate and price data on a daily basis
for approximately 400 Treasury securites. This function returns all that data
from 2010 to the last-completed business day.
}
\details{
The columns of the data.frame returned
\itemize{
    \item \bold{CUSIP}  Committee on Uniform Security Identification
    Procedures' nine-character, alphanumeric security identification code
    \item \bold{SECURITY.TYPE}
       \itemize{
           \item MARKET BASED BILL
           \item MARKET BASED NOTE
           \item MARKET BASED BOND
           \item TIPS
           \item MARKET BASED FRN
           }
    \item \bold{RATE}
    \item \bold{MATURITY.DATE}
    \item \bold{CALL.DATE} All \code{NA}
    \item \bold{BUY}
    \item \bold{SELL}
    \item \bold{END.OF.DAY}
    \item \bold{Date} Date for which the data was retrieved
    }
}
\examples{
library(ustreasuries)
fedinvest_data <- FedInvestData()
head(fedinvest_data)
tail(fedinvest_data)


}
\author{
George Fisher
}
\references{
US Dept. of the Treasury FedInvest Historical Prices
\url{https://www.treasurydirect.gov/GA-FI/FedInvest/selectSecurityPriceDate.htm}
}

