% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{ForwardPrice}
\alias{ForwardPrice}
\title{Forward price}
\usage{
ForwardPrice(Spot, Time, Interest, Yield, Income)
}
\arguments{
\item{Spot}{S_0, the spot price of the asset}

\item{Time}{T, the time to maturity}

\item{Interest}{r, the risk-free rate}

\item{Yield}{q, asset yield with continuous compounding}

\item{Income}{I, the PV of an asset's income}
}
\value{
the forward price
}
\description{
The delivery price in a forward contract that causes the contract to be worth zero.
}
\examples{
library(ustreasuries)
# Hull 7th edition Ch 5 P 103
Spot     <- 40
Time     <- 0.25
Interest <- 0.05
Yield    <- 0
Income   <- 0
ForwardPrice(Spot, Time, Interest, Yield, Income)

# Hull 7th edition Ch 5 P 105
Spot     <- 900
Time     <- 0.75
Interest <- 0.04
Yield    <- 0
Income   <- 40 * exp(-0.03 * 4/12) # PV(40) = 39.60
ForwardPrice(Spot, Time, Interest, Yield, Income)

# Hull 7th edition Ch 5 P 107
Spot     <- 25
Time     <- 0.50
Interest <- 0.10

# convert 0.04 discrete to continuous
Yield_d  <- 0.04
Yield    <- r_continuous(Yield_d, 2)

Income   <- 0
ForwardPrice(Spot, Time, Interest, Yield, Income)

}
\references{
Hull 7th edition Ch 5 P 103-108
}

