% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{ForwardRate}
\alias{ForwardRate}
\title{Forward rate}
\usage{
ForwardRate(SpotInterest1, Time1, SpotInterest2, Time2)
}
\arguments{
\item{SpotInterest1}{the zero rate for Time 1}

\item{Time1}{starting time}

\item{SpotInterest2}{the zero rate for Time 2}

\item{Time2}{ending time}
}
\value{
forward rate of inteest
}
\description{
The rate of interest for a period of time in the future implied by today's zero rates.
(discrete compounding)
}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

