% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{Gamma}
\alias{Gamma}
\title{Gamma of a European Option; convexity, curvature}
\usage{
Gamma(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Gamma of the option
}
\description{
Gamma describes the rate of change in the Delta; it is the second derivative of the option or portfolio of options with respect to the underlying asset's price.
}
\details{
If \code{abs(Gamma)} is large then the Delta is very sensitive to changes in the price of the underlying asset.
}
\note{
This Gamma masks that of package:stats; to use the latter write stats::Gamma()
}
\examples{
# Hull, 7th edition Ch 17 p 371,2
library(ustreasuries)
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

gamma <- Gamma(Stock, Exercise, Time, Interest, Yield, sigma)
round(gamma, 3) # 0.066

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition Ch 17 p 369-373
}

