% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{interest_rate}
\alias{interest_rate}
\title{Calculate annualized interest rate r(t, T) from a discount factor Z(t, T)}
\usage{
interest_rate(d_f, years, pmt_freq = Inf)
}
\arguments{
\item{d_f}{discount factor Z(t, T)}

\item{years}{number of (fractional) years: period that the discount factor
effects}

\item{pmt_freq}{frequency of compounding (optional)
\itemize{
    \item numeric => discrete compounding (2 => semi annual)
    \item Inf => continuous compounding (default)
    }}
}
\value{
r(t, T)
}
\description{
Calculate annualized interest rate r(t, T) from a discount factor Z(t, T)
}
\examples{
PV    <- 9000
FV    <- 13000
years <- 3
freq  <- 2   # compounding frequency = 2 => semi-annual

# continuous interest rate
# ========================
(r_continuous  <- CAGR(9000, 13000, years, type="continuous"))
(df_continuous <- discount_factor(r_continuous,  years))
(c_ir          <- interest_rate(df_continuous, years))
all.equal(r_continuous, c_ir)

# discrete interest rate
# ======================
(r_discrete  <- r_discrete(r_continuous, freq))
(df_discrete <- discount_factor(r_discrete,  years, freq))
(d_ir        <- interest_rate(df_discrete,   years, freq))
all.equal(r_discrete,   d_ir)

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Veronesi Ch2 P29-38
}

