% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{InTheMoneyCall}
\alias{InTheMoneyCall}
\title{Is a call option in the money?}
\usage{
InTheMoneyCall(Stock, Exercise)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}
}
\value{
IntrinsicValueCall(Stock, Exercise) > 0
}
\description{
Is the current market price of the asset above the strike price of the call?
}
\examples{
# http://www.call-options.com/in-the-money.html
library(ustreasuries)
Stock    <- 37.75     # S_0
Exercise <- 35        # K

InTheMoneyCall(Stock, Exercise)  # TRUE

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

