% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{InTheMoneyPut}
\alias{InTheMoneyPut}
\title{Is a put option in the money?}
\usage{
InTheMoneyPut(Stock, Exercise)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}
}
\value{
IntrinsicValuePut(Stock, Exercise) > 0
}
\description{
Is the current market price of the asset below the strike price of the put?
}
\examples{
# http://www.call-options.com/in-the-money.html
library(ustreasuries)
Stock    <- 35        # S_0
Exercise <- 37.50     # K

InTheMoneyPut(Stock, Exercise)  # TRUE

}

