% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{IntrinsicValueCall}
\alias{IntrinsicValueCall}
\title{Intrinsic value of a call option}
\usage{
IntrinsicValueCall(Stock, Exercise)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}
}
\value{
max(Stock - Exercise, 0)
}
\description{
The in-the-money portion of the option's premium
}
\examples{
# Investopia: Intrinsic Value
# http://www.investopedia.com/terms/i/intrinsicvalue.asp
Stock    <- 25 # S_0
Exercise <- 15 # K
IntrinsicValueCall(Stock, Exercise) # 10

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

