% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{IntrinsicValuePut}
\alias{IntrinsicValuePut}
\title{Intrinsic value of a put option}
\usage{
IntrinsicValuePut(Stock, Exercise)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}
}
\description{
The in-the-money portion of the option's premium
}
\examples{
# Investopia: Intrinsic Value
# http://www.investopedia.com/terms/i/intrinsicvalue.asp
Stock    <- 15 # S_0
Exercise <- 25 # K
IntrinsicValuePut(Stock, Exercise) # 10

}

