% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nelson.Siegel.R
\name{Nelson.Siegel.plot}
\alias{Nelson.Siegel.plot}
\title{Plot the actual and Nelson.Siegel Yield Curves}
\usage{
Nelson.Siegel.plot(actual_Rates, NS_rates, rows = 1,
  title = "Fitted Nelson-Siegel yield curve")
}
\arguments{
\item{actual_Rates}{a data.frame with the actual rates to be used in analysis}

\item{NS_rates}{The resulting Nelson Seigel rates}

\item{rows}{The rows in the two input matrices to plot}

\item{title}{plot title}
}
\value{
plot
}
\description{
Takes actual rates and Nelson Seigel rates and plot them
}
\examples{
library(ustreasuries)

# ======================================================
# example of plotting CMT data fitted with Nelson.Seigel
# ======================================================
all_data <- CMTrates()

# pick a selection of data around the date of the curve you want
rate.CMT    <- dplyr::filter(all_data,
                               all_data$NEW_DATE>=Sys.Date()-10 &
                               all_data$NEW_DATE<=Sys.Date())

NSParameters <- Nelson.Siegel(rate=rate.CMT)

Nelson.Siegel.plot(rate.CMT, NSrates(NSParameters), rows=5)

}
\author{
George Fisher
}
\seealso{
CMTrates, Nelson.Siegel, NSrates
}

