% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nelson.Siegel.R
\name{Nelson.Siegel}
\alias{Nelson.Siegel}
\title{Estimation of the Nelson-Siegel parameters}
\usage{
Nelson.Siegel(rate, maturity = c(1/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30))
}
\arguments{
\item{rate}{vector or matrix which contains the interest rates}

\item{maturity}{vector wich contains the maturity ( in months) of the rate.
The vector's length must be the same of the number of columns of the rate.}
}
\value{
Returns a data frame with the estimated coefficients:
\eqn{\beta_{0t}, \beta_{1t}, \beta_{2t},  \lambda}
}
\description{
Returns the estimated coefficients of the Nelson-Siegel's model.
}
\examples{
library(ustreasuries)

# ==============================================
# example of fitting CMT data with Nelson.Seigel
# ==============================================
all_data <- CMTrates()

# pick a selection of data around the date of the curve you want
rate.CMT    <- dplyr::filter(all_data,
                               all_data$NEW_DATE>=Sys.Date()-10 &
                               all_data$NEW_DATE<=Sys.Date())

NSParameters <- Nelson.Siegel(rate=rate.CMT)

head(NSParameters)

Nelson.Siegel.plot(rate.CMT, NSrates(NSParameters), rows=5)

}
\author{
From \code{YieldCurve} package by Sergio Salvino Guirreri

Modified by George Fisher
}

