% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{nprime}
\alias{Nprime}
\alias{enprime}
\alias{nprime}
\title{N': the first derivative of N(x) ... the Standard Normal PDF}
\usage{
nprime(x)
}
\arguments{
\item{x}{real number}
}
\value{
N'(x)
}
\description{
Equivalent to the R function \emph{dnorm}
}
\note{
same function provided by \emph{nprime, Nprime, enprime}
}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

