% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nelson.Siegel.R
\name{NSrates}
\alias{NSrates}
\title{Interest rates of the Nelson-Siegel model}
\usage{
NSrates(Coeff, maturity = c(1/12, 3/12, 6/12, 1, 2, 3, 5, 7, 10, 20, 30))
}
\arguments{
\item{Coeff}{Vector or matrix of the beta's coefficients and lambda as the function Nelson.Siegel returns.}

\item{maturity}{maturity of the yield curve of which want to return the interest rates.}
}
\value{
Return interest rates in matrix object with number of rows equal to nrow(betaCoeff) and number of columns equal to length(maturity).
}
\description{
Returns the interest rates by Nelson-Siegel's model.
}
\details{
Coeff is a vector or matrix of the four coefficients of the Nelson-Siegel's model:
\eqn{\beta_0, \beta_1, \beta_2,  \lambda}
}
\examples{
library(ustreasuries)
all_data <- CMTrates()

# pick a selection of data around the date of the curve you want
rate.CMT    <- dplyr::filter(all_data,
                               all_data$NEW_DATE>=Sys.Date()-10 &
                               all_data$NEW_DATE<=Sys.Date())

NSParameters <- Nelson.Siegel(rate=rate.CMT)

head(NSrates(NSParameters))

}
\author{
From \code{YieldCurve} package by Sergio Salvino Guirreri

Modified by George Fisher
}

