% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nelson.Siegel.R
\name{NSzeros.plot}
\alias{NSzeros.plot}
\title{Plot Z(0, T) and r(0, T)}
\usage{
NSzeros.plot(r, Z)
}
\arguments{
\item{r}{an xts list of NSrates}

\item{Z}{a data.frame list of NSzeros}
}
\value{
a plot
}
\description{
Plot the discount factors and fitted rates derived from the NSrates and NSzeros functions
}
\examples{
library(ustreasuries)
all_data <- CMTrates()

rate.CMT    <- dplyr::filter(all_data,
                        all_data$NEW_DATE>=Sys.Date()-10 &
                        all_data$NEW_DATE<=Sys.Date())
NSParameters <- Nelson.Siegel(rate=rate.CMT)

r <- NSrates(NSParameters)[1,]           # r(0, T)
Z <- NSzeros(NSrates(NSParameters))[1,]  # Z(0, T)

NSzeros.plot(r, Z)

}
\author{
George Fisher
}

