% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nelson.Siegel.R
\name{NSzeros}
\alias{NSzeros}
\title{Convert NSrates to discount factors Z(0, T)}
\usage{
NSzeros(NS_rates)
}
\arguments{
\item{NS_rates}{the value returned by the \code{NSrates()} function}
}
\value{
\code{data.frame} rownames are the dates,
colnames are the time to maturity in character format preceded by "X"
}
\description{
Returns the discount factors associated with Nelson Seigel fitted rates
}
\details{
\eqn{Z(0, T) = e^{-r(0, T) * T}}
}
\examples{
library(ustreasuries)
all_data <- CMTrates()

rate.CMT    <- dplyr::filter(all_data,
                               all_data$NEW_DATE>=Sys.Date()-10 &
                               all_data$NEW_DATE<=Sys.Date())

NSParameters <- Nelson.Siegel(rate=rate.CMT)

head(NSzeros(NSrates(NSParameters)))

}
\author{
George Fisher
}
\references{
Veronesi, P., \emph{Fixed Income Securities} p67
}

