% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{phi}
\alias{phi}
\title{Numerical Derivation of the Standard Normal CDF}
\usage{
phi(x)
}
\arguments{
\item{x}{real number}
}
\value{
N(x)
}
\description{
Equivalent to the R function \emph{pnorm}
}

