% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondMath.R
\name{Price}
\alias{Price}
\title{Bond Price}
\usage{
Price(ytm, coupon, years)
}
\arguments{
\item{ytm}{yield to maturity}

\item{coupon}{the annual coupon rate}

\item{years}{Fractional years: 2 years, 3 months, 10 days = 2 + 3/12 + 10/365 = 2.277397}
}
\value{
bond price
}
\description{
"Lazy-Man's" bond price: no dates, just fractional years
}
\examples{
# Veronesi example 2.11
library(ustreasuries)

Price(0.0375479295223684, 0.04750, 9.5)
Price(0.0366032177811388, 0.08875, 9.5)

}
\author{
George Fisher
}

