% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintYieldCurves.R
\name{PrintYieldCurves}
\alias{PrintYieldCurves}
\title{Print yield curves}
\usage{
PrintYieldCurves(YieldCurveDF, rows = c(1), title = "Yield Curve",
  tabprt = FALSE)
}
\arguments{
\item{YieldCurveDF}{data.frame created by the \emph{CMTrates} function}

\item{rows}{a vector of integer row indices indicating the row(s) to use
for plotting}

\item{title}{a string to use as the title of the plot}

\item{tabprt}{if TRUE, print a table of the rate data used}
}
\value{
nothing: produces a plot and optionally a table
}
\description{
Given a data.frame returned by the \emph{CMTrates}, function, you can
specify one or more dates to use to print the yield curve on those dates.
}
\examples{
library(ustreasuries)
all_data <- CMTrates()
PrintYieldCurves(dplyr::filter(all_data,
                                   all_data$NEW_DATE>=as.Date("2006-01-01") &
                                   all_data$NEW_DATE<=as.Date("2009-12-31")),
                   rows=c(1, 272, 272*2, 272*3),
                   title="Yield Curves Before & During the Financial Crisis",
                   tabprt=TRUE)

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

