% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{PutParity}
\alias{PutParity}
\title{Put prices derived from put-call parity}
\usage{
PutParity(Stock, Exercise, Time, Interest, Yield, Call_price)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{Call_price}{the price of the call option to convert}
}
\description{
Put prices derived from put-call parity
}
\examples{
# Hull 7th edition Ch 17 P 357
library(ustreasuries)
Stock    <- 49      # S_0
Exercise <- 50      # K
Time     <- 20/52   # T
Interest <- 0.05    # r
Yield    <- 0.13#0  # q
sigma    <- 0.20

EC = EuroCall(Stock, Exercise, Time, Interest, Yield, sigma)
EP = EuroPut(Stock, Exercise, Time, Interest, Yield, sigma)

PC = CallParity(Stock, Exercise, Time, Interest, Yield, EP)
PP = PutParity(Stock, Exercise, Time, Interest, Yield, EC)

writeLines(paste0("European Call Price: ", EC, "\\n",
                  "Call Parity Price:   ", PC, "\\n",
                  "Difference:          ", EC-PC, "\\n\\n",

                  "European Put Price:  ", EP, "\\n",
                  "Put Parity Price:    ", PP, "\\n",
                  "Difference:          ", EP-PP))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

