% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{r_continuous}
\alias{r_continuous}
\title{Convert TO continuous compounding FROM discrete}
\usage{
r_continuous(r_d, compounding_periods_per_year)
}
\arguments{
\item{r_d}{the discrete CAGR returned from \emph{CAGRd}}

\item{compounding_periods_per_year}{how often the rate is compounded (2 => semiannual)}
}
\value{
the continuously-compounded rate of return
}
\description{
Convert TO continuous compounding FROM discrete
}
\examples{
# Hull 7th edition Ch 5 P 107
r_d                          <- 0.04
compounding_periods_per_year <- 2
ans <- r_continuous(r_d, compounding_periods_per_year)
writeLines(paste0(round(ans,4)))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

