% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{r_discrete}
\alias{r_discrete}
\title{Convert TO discrete compounding FROM continuous}
\usage{
r_discrete(r_c, freq)
}
\arguments{
\item{r_c}{the continuously compounded rate of return}

\item{freq}{how often the discrete rate is compounded (2 => semiannual)}
}
\value{
the discrete rate of return
}
\description{
Convert TO discrete compounding FROM continuous
}
\note{
FV = PV * (1 + discrete / freq) ** (freq * years)

FV = PV * exp(continuous * years)
}
\examples{
PV    <- 9000
FV    <- 13000
years <- 3
freq  <- 2   # compounding frequency = 2 => semi-annual

(r_continuous <- CAGR(PV, FV, years, type="continuous"))
(r_discrete   <- r_discrete(r_continuous, freq))

PV * (1 + r_discrete / freq) ** (freq * years)

PV * exp(r_continuous * years)

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

