% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{RandN}
\alias{RandN}
\title{Produces a standard normal random variable 'epsilon'}
\usage{
RandN()
}
\value{
epsilon in N(0, 1)
}
\description{
Random number from a Gausiian distribution
}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

