% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{RandN_ssdt}
\alias{RandN_ssdt}
\title{Produces a standard normal random normal variable epsilon times sigma*sqrt(deltaT)}
\usage{
RandN_ssdt(ssdt)
}
\arguments{
\item{ssdt}{variance}
}
\value{
epsilon in N(0, ssdt)
}
\description{
random number from a Gausiian distribution with variance ssdt
}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}

