% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{RhoCall}
\alias{RhoCall}
\title{Rho of a European Call Option}
\usage{
RhoCall(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Rho of the call option
}
\description{
Rho is the sensitivity of an option price to changes in the risk-free
rate of interest
}
\note{
for futures and foreign exchange, use the specialized functions
RhoFuturesCall, RhoFuturesPut, RhoFXCall, RhoFXPut
}
\examples{
# Hull, 7th edition Ch 17 p 376
library(ustreasuries)
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

rho <- RhoCall(Stock, Exercise, Time, Interest, Yield, sigma)
round(rho, 2) # 8.91

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition ch 17 p375-376
}

