% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{RhoPut}
\alias{RhoPut}
\title{Rho of a European Put Option}
\usage{
RhoPut(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Rho of the put option
}
\description{
Rho is the sensitivity of an option price to changes in the risk-free
rate of interest
}
\note{
for futures and foreign exchange, use the specialized functions
RhoFuturesCall, RhoFuturesPut, RhoFXCall, RhoFXPut
}
\examples{
# Hull, 7th edition Ch 17 p 376
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

rhocall <- RhoCall(Stock, Exercise, Time, Interest, Yield, sigma)
rhoput  <- RhoPut(Stock, Exercise, Time, Interest, Yield, sigma)

writeLines(paste0("RhoCall:  ", round(rhocall, 2), "\\n",
                  "RhoPut:  ", round(rhoput, 2)))

}
\references{
Hull, 7th edition ch 17 p375-376
}

