% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rates.R
\name{SP500}
\alias{SP500}
\title{Daily S&P 500 data from Jan 3, 1950 to present}
\usage{
SP500()
}
\value{
A data.frame with the daily data
}
\description{
The "usual" S&P 500 index, without dividends reinvested.
}
\details{
The columns of the data.frame returned
\itemize{
    \item \bold{Date}
    \item \bold{Open}
    \item \bold{High}
    \item \bold{Low}
    \item \bold{Close}
    \item \bold{Volume}
    \item \bold{Adj.Close}
    }
}
\examples{
sp500_idx <- SP500()
head(sp500_idx)
tail(sp500_idx)

}
\author{
George Fisher
}
\references{
Yahoo Finance
}

