% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rates.R
\name{SP500TR}
\alias{SP500TR}
\title{Daily S&P 500 Total Return data from Jan 4, 1988 to present}
\usage{
SP500TR()
}
\value{
A data.frame with the daily data
}
\description{
The S&P 500 index, with dividends reinvested.
}
\details{
The columns of the data.frame returned
\itemize{
    \item \bold{Date}
    \item \bold{Open}
    \item \bold{High}
    \item \bold{Low}
    \item \bold{Close}
    \item \bold{Volume}
    \item \bold{Adj.Close}
    }
}
\examples{
sp500_idx_tr <- SP500TR()
head(sp500_idx_tr)
tail(sp500_idx_tr)

}
\author{
George Fisher
}
\references{
Yahoo Finance
}

