% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{ThetaCall}
\alias{ThetaCall}
\title{Theta of a European Call Option}
\usage{
ThetaCall(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
The Theta of the call option
}
\description{
Theta is the decay in the value of an option or a portfolio of options as time passes
}
\details{
In a delta-neutral portfolio, Theta is a proxy for Gamma
}
\note{
divide by 365 for "per calendar day"; 252 for "per trading day"
}
\examples{
# Hull, 7th edition Ch 17 p 367
library(ustreasuries)
Stock    <- 49     # S_0
Exercise <- 50     # K
Time     <- 20/52  # T
Interest <- 0.05   # r
Yield    <- 0      # q
sigma    <- 0.20

thcall <- ThetaCall(Stock, Exercise, Time, Interest, Yield, sigma)

writeLines(paste0("Theta:            ", round(thcall, 2),     "\\n",
                  "per calendar day: ", round(thcall/365, 4), "\\n",
                  "per trading day:  ", round(thcall/252, 4)))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition ch 17 p367-368
}

