% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{TimeValueCall}
\alias{TimeValueCall}
\title{Time Value of a European Call Option}
\usage{
TimeValueCall(Stock, Exercise, Call_price)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}

\item{Call_price}{The value of the call being analyzed}
}
\value{
The Time Value of the call option
}
\description{
The total value of an option is its intrinsic value plus its time value
}
\examples{
# Hull, 7th edition ch 8 p186
Stock      <- 21     # S_0
Exercise   <- 20     # K
Call_price <- 1.875

(calltv <- TimeValueCall(Stock, Exercise, Call_price))

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull, 7th edition ch 8 p186
}

