% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{TimeValuePut}
\alias{TimeValuePut}
\title{Time Value of a European Put Option}
\usage{
TimeValuePut(Stock, Exercise, Put_price)
}
\arguments{
\item{Stock}{S_0, the asset price}

\item{Exercise}{K, the option strike price}

\item{Put_price}{The value of the put being analyzed}
}
\value{
The Time Value of the put option
}
\description{
The total value of an option is its intrinsic value plus its time value
}
\examples{
# Hull, 7th edition ch 8 p186
Stock      <- 21     # S_0
Exercise   <- 20     # K
Put_price <- 1.875

(puttv <- TimeValuePut(Stock, Exercise, Put_price))

}
\references{
Hull, 7th edition ch 8 p186
}

