% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondMath.R
\name{YearsMonthsDays}
\alias{YearsMonthsDays}
\title{Convert fractional years into a lubridate period object}
\usage{
YearsMonthsDays(years)
}
\arguments{
\item{years}{Fractional years: 2 years, 3 months, 10 days = 2 + 3/12 + 10/365 = 2.277397}
}
\value{
a lubridate period object
}
\description{
Given a fractional number of years you need to convert into years, months and
days to do date math

See the example for the conversion
}
\examples{
library(ustreasuries)
(diff_tm <- YearsMonthsDays(2 + 3/12 + 10/365))

# July 1, 2000 plus 2 years (2002), 3 months (10), 10 days (11)
as.Date(lubridate::ymd("2000/07/01") + diff_tm)

}
\author{
George Fisher
}

