% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondMath.R
\name{YTM}
\alias{YTM}
\title{Yield To Maturity}
\usage{
YTM(price, coupon, years)
}
\arguments{
\item{price}{the actual price paid}

\item{coupon}{the annual coupon rate}

\item{years}{Fractional years: 2 years, 3 months, 10 days = 2 + 3/12 + 10/365 = 2.277397}
}
\value{
the YTM
}
\description{
"Lazy-Man's" YTM: no dates, just fractional years
}
\examples{
# Veronesi example 2.11
library(ustreasuries)

ytm <- YTM(107.8906, 0.0475, 9.5)
print(paste0(ytm, " = ", round(ytm*100,4),"\%"))
ytm <- YTM(141.5267, 0.08875, 9.5)
print(paste0(ytm, " = ", round(ytm*100,4),"\%"))

}
\author{
George Fisher
}

