% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BondMath.R
\name{Zbootstrap}
\alias{Zbootstrap}
\title{Use the Booststrap Method to determine discount factors Z(0, T_i)}
\usage{
Zbootstrap(prices, coupons)
}
\arguments{
\item{prices}{a vector of prices from the shortest (6-months) to the longest maturity}

\item{coupons}{a vector, usually beginning with 0, with the annual coupon
rate as a decimal}
}
\value{
Z a vector of the discount factors
}
\description{
Assuming a sequence of semi-annual coupon bonds beginning with
a maturity equal to six months and increasing, this function determines
the sequence of discount factors for each time segment using the boothstrap
method. This method is limited by the facts that it assumes evenly-spaced
six-month maturities and an unbroken sequence. It is also iterative instead
of using matrix algebra.
}
\examples{
# Veronesi Ch2 p 46
p1 <- 98.3607
c1 <- 0

p2 <- 99.2343
c2 <- 0.0275

p3 <- 99.1093
c3 <- 0.03

prices  <- c(p1, p2, p3)
coupons <- c(c1, c2, c3)

Zbootstrap(prices, coupons)

}
\author{
George Fisher
}
\references{
Veronesi Ch2 p46-47
}
\seealso{
Nelson Seigel and Svensson are more helpful for dirty, real-life bond data.
}

