% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{discount_factor}
\alias{discount_factor}
\title{Calculate discount factor Z(t, T)}
\usage{
discount_factor(annual_rate, years, pmt_freq = Inf)
}
\arguments{
\item{annual_rate}{the compounded annual rate of interest}

\item{years}{number of (fractional) years}

\item{pmt_freq}{frequency of compounding (optional)
\itemize{
    \item numeric => discrete compounding (2 => semi annual)
    \item Inf => continuous compounding (default)
    }}
}
\value{
Z(t, T)
}
\description{
Calculate discount factor Z(t, T)
}
\examples{
PV    <- 9000
FV    <- 13000
years <- 3
freq  <- 2   # compounding frequency = 2 => semi-annual

(continuous <- CAGR(PV, FV, years, type="continuous"))
(discrete   <- r_discrete(r_continuous, freq))

(df_continuous <- discount_factor(continuous,  years))

(df_discrete   <- discount_factor(discrete,  years, freq))

FV * df_continuous
FV * df_discrete

all.equal(df_continuous, df_discrete) # df_continuous == df_discrete

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Veronesi Ch2 P29-38
}

