% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM.R
\name{EuroPut}
\alias{EuroPut}
\title{European Put}
\usage{
EuroPut(Stock, Exercise, Time, Interest, Yield, sigma)
}
\arguments{
\item{Stock}{S0, the initial stock price}

\item{Exercise}{K, the strike price}

\item{Time}{T, the time to maturity in fractional years}

\item{Interest}{r, the risk-free rate of return}

\item{Yield}{q, the dividend yield}

\item{sigma}{the asset volatility}
}
\value{
the price of a European Put Option
}
\description{
A put option that can only be exercised at expiration
}
\examples{
# Hull 7th edition Ch 13 P 294
Stock     <- 42     # S_0
Exercise  <- 40     # K
Time      <- 0.50   # T
Interest  <- 0.10   # r
Yield     <- 0      # q
sigma     <- 0.20
ans <- EuroPut(Stock, Exercise, Time, Interest, Yield, sigma)
round(ans,2)        # 0.81

\dontrun{
GenBS(Stock, Exercise, Interest, sigma, Time, Yield)
}

}
\author{
George Fisher \email{GeorgeRFisher@gmail.com}
}
\references{
Hull 7th edition Ch 13 P 291
}

