% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rates.R
\name{NSzeros}
\alias{NSzeros}
\title{Convert rates to discount factors Z(0, T)}
\usage{
NSzeros(NS_rates)
}
\arguments{
\item{NS_rates}{the value returned by the \code{NSrates()} or \code{Srates()}functions}
}
\value{
\code{data.frame} rownames are the dates,
colnames are the time to maturity in character format preceded by "X"
}
\description{
Returns the discount factors associated with Nelson Seigel fitted rates
}
\details{
\eqn{Z(0, T) = e^{-r(0, T) * T}}
}
\author{
George Fisher
}
\references{
Veronesi, P., \emph{Fixed Income Securities} p67
}

