% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BSM_utilities.R
\name{r_discrete}
\alias{r_discrete}
\title{Convert TO discrete compounding FROM continuous}
\usage{
r_discrete(r_c, compounding_periods_per_year)
}
\arguments{
\item{r_c}{the continuous CAGR returned from \emph{r_continuous}}

\item{compounding_periods_per_year}{how often the rate is compounded (2 => semiannual)}
}
\value{
the discrete rate of return
}
\description{
r_discrete = m * (exp(r_continuous / m) - 1)
where m is the number of compounding periods per year
}
\examples{
# reverse the example from Hull 7th edition Ch 5 P 107
r_c                          <- 0.03960525
compounding_periods_per_year <- 2
ans <- r_discrete(r_c, compounding_periods_per_year)
writeLines(paste0(round(ans,4)))

}

